🎨 NavigationItemをクライアントコンポーネントとして切り出す
