+++ src/app/_components/Navigation/NavigationItem.jsx
@@ -1,0 +1,24 @@
+"use client";
+
+import Link from "next/link";
+import { usePathname } from "next/navigation";
+
+// ナビゲーションアイテム
+export const NavigationItem = ({ text, icon, href }) => {
+  console.log("clientです");
+  // hrefと現在のパスが一致している場合にtrueを返す
+  const isActive = usePathname() === href;
+
+  return (
+    <Link href={href} className="flex-1 select-none">
+      <div
+        className={`flex w-full flex-col items-center gap-y-0.5 hover:border-slate-100 sm:-ml-3 sm:w-fit sm:flex-row sm:gap-2 sm:rounded-full sm:py-1.5 sm:pl-3 sm:pr-5 sm:hover:bg-slate-100 sm:active:bg-slate-100 ${
+          isActive ? "font-bold" : ""
+        }`}
+      >
+        <div className="h-6 w-6">{icon}</div>
+        <div className="text-xs sm:text-base">{text}</div>
+      </div>
+    </Link>
+  );
+};
